requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-crispy-forms==1.6.1
idna==2.5
oauthlib==2.0.2
PyJWT==1.5.2
pyspark==2.1.1+hadoop2.7
python3-openid==3.1.0
pytz==2017.2
requests==2.18.2
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
